Created AddExpenseView, UI elements divided from AddExpenseViewController
