+++ Expense Tracker/Source/Presentation/Modules/AddExpense/AddExpenseViewController.swift
@@ -15,16 +15,7 @@ final class AddExpenseViewController: UIViewController,
     // MARK: - Private properties
     
     private let presenter: AddExpensePresenterProtocol
-    
-    private lazy var myButton: UIButton = {
-        let myButton = UIButton()
-        myButton.setTitle(Strings.signUpButtonTitle, for: .normal)
-        myButton.translatesAutoresizingMaskIntoConstraints = false
-        myButton.backgroundColor = .systemBlue
-        myButton.setTitleColor(.white, for: .normal)
-        myButton.addTarget(self, action: #selector(buttonTapped), for: .touchUpInside)
-        return myButton
-    }()
+    private let addExpenseView = AddExpenseView()
     
     // MARK: - Init
     
@@ -36,25 +27,17 @@ final class AddExpenseViewController: UIViewController,
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-    
+
     // MARK: - Life Cycle
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        view.addSubview(myButton)
-        setupAutoLayout()
+        view.addSubview(addExpenseView)
+        addExpenseView.frame = view.bounds
+        addExpenseView.myButton.addTarget(self, action: #selector(buttonTapped), for: .touchUpInside)
     }
     
     // MARK: - Private func
-    
-    private func setupAutoLayout() {
-        NSLayoutConstraint.activate([
-            myButton.centerXAnchor.constraint(equalTo: view.centerXAnchor),
-            myButton.centerYAnchor.constraint(equalTo: view.centerYAnchor),
-            myButton.widthAnchor.constraint(equalToConstant: 200),
-            myButton.heightAnchor.constraint(equalToConstant: 40)
-        ])
-    }
     
     @objc private func buttonTapped() {
         print("Button tapped!")
